refactor: components to index (barrel) file
