+++ src/components/Footer/footer.tsx
@@ -52,4 +52,4 @@ function Footer() {
   );
 }
 
-export default Footer;
+export { Footer };
