+++ src/components/index.ts
@@ -1,0 +1,6 @@
+export { BasePage } from "./BasePage/BasePage";
+export { Footer } from "./Footer/footer";
+export { NavBar } from "./NavBar/navBar";
+export { SEO } from "./SEO/seo";
+export { SponsorCard } from "./SponsorCard/sponsorCard";
+export { SponsorSection } from "./SponsorSection/sponsorSection";
