+++ src/pages/calendar.tsx
@@ -1,8 +1,7 @@
 import React from "react";
 import { HeadFC } from "gatsby";
 
-import BasePage from "../components/BasePage/BasePage";
-import { SEO } from "../components/SEO/seo";
+import { BasePage, SEO } from "../components";
 
 const calendarUrl="https://calendar.google.com/calendar/embed?height=600&wkst=1&bgcolor=%23ff007f&ctz=America%2FDenver&title=Blazer%20Bots%20Calendar&src=bDhlajBrc2R1OTN0ZmRsZGJhdDB2MWMycjEzODQ0Y3RAaW1wb3J0LmNhbGVuZGFyLmdvb2dsZS5jb20&src=b2hzYmxhemVyYm90c0BnbWFpbC5jb20&color=%23E67C73&color=%23D50000"
